feat(refactor): 임베딩 관련 공통 로직 이동
