+++ backend/services/base/embedder.py
@@ -1,14 +1,30 @@
 from abc import abstractmethod
-from typing import Generic, List, TypeVar
+from typing import Dict, Generic, List, Optional, TypeVar
 
 from aiohttp import ClientSession
+import requests
 
+from mixins.asyncio import retry_async
+from mixins.http_client import HTTPMetaclass
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+EMBED_URL = os.environ.get("EMBED_URL")
+
+
+class EmbedResult(Dict):
+    chunk: Optional[str] = None
+    dense: List[float]
+    sparse: Dict[int, float]
 
 
 DTO = TypeVar("DTO")
 
 
 class BaseEmbedder(Generic[DTO], metaclass=HTTPMetaclass):
+
     async def embed_dto_async(
         self, item: DTO, session: Optional[ClientSession] = None
     ) -> DTO:
@@ -38,3 +54,31 @@ class BaseEmbedder(Generic[DTO], metaclass=HTTPMetaclass):
     ) -> List[DTO]:
         pass
 
+    @retry_async(delay=3)
+    async def _embed_async(
+        self,
+        texts: str | List[str],
+        session: ClientSession,
+        chunking: bool = True,
+        truncate: bool = True,
+    ) -> EmbedResult | List[EmbedResult]:
+        body = {"inputs": texts, "chunking": chunking, "truncate": truncate}
+        async with session.post(f"{EMBED_URL}/embed", json=body) as res:
+            if res.status == 200:
+                data = await res.json()
+                return data
+
+            raise Exception("텍스트 임베딩에 실패했습니다.")
+
+    def _embed_query(
+        self,
+        query: str | List[str],
+        chunking: bool = True,
+        truncate: bool = True
+    ):
+        body = {"inputs": query, "chunking": chunking, "truncate": truncate}
+        res = requests.post(f"{EMBED_URL}/embed", json=body)
+        if res.status_code == 200:
+            return res.json()
+
+        raise Exception
